Added ExerciseEntityCategory data model and changed codegen for models to only manually create the class and moved custom properties to the class file to allow automatic generation when changing the data models and relationships.
