--- Iron/CoreData/DataController.swift
@@ -1,24 +1,0 @@
-//
-//  DataController.swift
-//  Iron
-//
-//  Created by Nick Schwab on 6/1/22.
-//
-
-import CoreData
-import Foundation
-
-class DataController: ObservableObject {
-    let container = NSPersistentContainer(name: "Iron")
-    
-    init() {
-        container.loadPersistentStores { description, error in
-            if let error = error {
-                print("Core data failed to load: \(error.localizedDescription)")
-                return
-            }
-            self.container.viewContext.mergePolicy = NSMergePolicy.mergeByPropertyObjectTrump
-        }
-    }
-    
-}
