+++ Iron/CoreData/ESet+CoreDataClass.swift
@@ -13,3 +13,22 @@ import CoreData
 public class ESet: NSManagedObject {
 
 }
+extension ESet {
+    
+    public var strRir: String {
+        get {rir ?? "11"}
+        set {self.rir = newValue}
+    }
+    public var strRpe: String {
+        get {rpe ?? "11"}
+        set {self.rpe = newValue}
+    }
+    public var strWeight: String {
+        get {weight ?? "111.11"}
+        set {self.weight = newValue}
+    }
+    public var strReps: String {
+        get { reps ?? "11"}
+        set {self.reps = newValue}
+    }
+}
