+++ Iron/CoreData/ExerciseEntityCategory+CoreDataClass.swift
@@ -1,0 +1,15 @@
+//
+//  ExerciseEntityCategory+CoreDataClass.swift
+//  Iron
+//
+//  Created by Nick Schwab on 7/1/22.
+//
+//
+
+import Foundation
+import CoreData
+
+@objc(ExerciseEntityCategory)
+public class ExerciseEntityCategory: NSManagedObject {
+
+}
